+++ spec/untangled/client/core_spec.cljs
@@ -164,51 +164,63 @@
 
       (uc/mount* mounted-mock-app :fake-root :dom-id)))
   (behavior "When is is not already mounted"
-    (let [mock-app {:mounted? false :initial-state {:a 1} :reconciler-options :OPTIONS}]
-      (when-mocking
-        (uc/initialize app state root dom opts) => (do
-                                                     (assertions
-                                                       "Initializes the app with a plain map when root does not implement InitialAppState"
-                                                       state => {:a 1}
-                                                       ))
-
-        (uc/mount* mock-app :fake-root :dom-id)))
-    (let [supplied-atom (atom {:a 1})
-          mock-app {:mounted? false :initial-state supplied-atom :reconciler-options :OPTIONS}]
-      (when-mocking
-        (uc/initialize app state root dom opts) => (do
-                                                     (assertions
-                                                       "Initializes the app with a supplied atom when root does not implement InitialAppState"
-                                                       {:a 1} => @state))
-
-        (uc/mount* mock-app :fake-root :dom-id)))
-    (let [mock-app {:mounted? false :initial-state {:a 1} :reconciler-options :OPTIONS}]
-      (when-mocking
-        (log/warn msg) =1x=> (do (assertions "warns about duplicate initialization"
-                                   msg =fn=> (partial re-matches #"^You supplied.*")))
-        (uc/initialize app state root dom opts) => (do
-                                                     (assertions
-                                                       "Initializes the app with the InitialAppState"
-                                                       state => (uc/initial-state Parent nil)))
-
-        (uc/mount* mock-app Parent :dom-id)))
-    (let [mock-app {:mounted? false :initial-state (atom {:a 1}) :reconciler-options :OPTIONS}]
-      (behavior "When both atom and InitialAppState are present:"
+    (behavior "and root does NOT implement InitialAppState"
+      (let [mock-app {:mounted? false :initial-state {:a 1} :reconciler-options :OPTIONS}]
         (when-mocking
-          (log/warn msg) =1x=> true
-          (om/tree->db c d merge-idents) => (do
-                                              (behavior "Normalizes InitialAppState:"
-                                                (assertions
-                                                  "includes Om tables"
-                                                  merge-idents => true
-                                                  "uses the Root UI component query"
-                                                  c => Parent
-                                                  "uses InitialAppState as the data"
-                                                  d => (uc/initial-state Parent nil)))
-                                              :NORMALIZED-STATE)
           (uc/initialize app state root dom opts) => (do
                                                        (assertions
-                                                         "Overwrites the supplied atom with the normalized InitialAppState"
-                                                         @state => :NORMALIZED-STATE))
+                                                         "Initializes the app with a plain map"
+                                                         state => {:a 1}
+                                                         ))
 
-          (uc/mount* mock-app Parent :dom-id))))))
+          (uc/mount* mock-app :fake-root :dom-id)))
+      (let [supplied-atom (atom {:a 1})
+            mock-app {:mounted? false :initial-state supplied-atom :reconciler-options :OPTIONS}]
+        (when-mocking
+          (uc/initialize app state root dom opts) => (do
+                                                       (assertions
+                                                         "Initializes the app with a supplied atom"
+                                                         {:a 1} => @state))
+
+          (uc/mount* mock-app :fake-root :dom-id))))
+    (behavior "and root IMPLEMENTS InitialAppState"
+      (let [mock-app {:mounted? false :initial-state {:a 1} :reconciler-options :OPTIONS}]
+        (when-mocking
+          (log/warn msg) =1x=> (do (assertions "warns about duplicate initialization"
+                                     msg =fn=> (partial re-matches #"^You supplied.*")))
+          (uc/initialize app state root dom opts) => (do
+                                                       (assertions
+                                                         "Initializes the app with the InitialAppState"
+                                                         state => (uc/initial-state Parent nil)))
+
+          (uc/mount* mock-app Parent :dom-id)))
+      (let [mock-app {:mounted? false :initial-state (atom {:a 1}) :reconciler-options :OPTIONS}]
+        (behavior "When both atom and InitialAppState are present:"
+          (when-mocking
+            (log/warn msg) =1x=> true
+            (om/tree->db c d merge-idents) => (do
+                                                (behavior "Normalizes InitialAppState:"
+                                                  (assertions
+                                                    "includes Om tables"
+                                                    merge-idents => true
+                                                    "uses the Root UI component query"
+                                                    c => Parent
+                                                    "uses InitialAppState as the data"
+                                                    d => (uc/initial-state Parent nil)))
+                                                :NORMALIZED-STATE)
+            (uc/initialize app state root dom opts) => (do
+                                                         (assertions
+                                                           "Overwrites the supplied atom with the normalized InitialAppState"
+                                                           @state => :NORMALIZED-STATE))
+
+            (uc/mount* mock-app Parent :dom-id))))
+      (let [mock-app {:mounted? false :reconciler-options :OPTIONS}]
+        (behavior "When only InitialAppState is present:"
+          (when-mocking
+            (untangled.client.core/initial-state root-component nil) => :INITIAL-UI-STATE
+            (uc/initialize app state root dom opts) => (do
+                                                         (assertions
+                                                           "Supplies the raw InitialAppState to internal initialize"
+                                                           state => :INITIAL-UI-STATE))
+
+            (uc/mount* mock-app Parent :dom-id)))))))