+++ src/untangled/client/impl/data_fetch.cljs
@@ -16,7 +16,7 @@
 
 ; TODO: Some of this API is public, and should be in the non-impl ns.
 
-(declare data-path data-uuid data-query set-loading! full-query loaded-callback error-callback data-marker?)
+(declare data-target data-path data-uuid data-query set-loading! full-query loaded-callback error-callback data-marker?)
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; Implementation for public api
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
@@ -82,7 +82,7 @@
    An element is a duplicate IFF (keys-fn element) has key collision with any prior element
    to come before it. E.g. (dedupe-by identity [[:a] [:b] [:a] [:a :c]]) => [[:a] [:b]]
    Returns a stateful transducer when no collection is provided."
-  ([keys-fn] ;; transducer fn
+  ([keys-fn]                                                ;; transducer fn
    (fn [rf]
      (let [keys-seen (volatile! #{})]
        (fn
@@ -210,7 +210,8 @@
 (defn ready-state
   "Generate a ready-to-load state with all of the necessary details to do
   remoting and merging."
-  [& {:keys [ident field params without query post-mutation fallback parallel refresh marker] :or {without #{} refresh [] marker true}}]
+  [{:keys [ident field params without query post-mutation post-mutation-params fallback parallel refresh marker target]
+    :or   {without #{} refresh [] marker true}}]
   (assert (or field query) "You must supply a query or a field/ident pair")
   (assert (or (not field) (and field (util/ident? ident))) "Field requires ident")
   (let [old-ast (om/query->ast query)
@@ -221,37 +222,31 @@
         key (if (util/join? query-field) (util/join-key query-field) query-field)
         query' (om/ast->query ast)]
     (assert (or (not field) (= field key)) "Component fetch query does not match supplied field.")
-    {::type          :ready
-     ::uuid          (uuid/uuid-string (uuid/make-random-squuid))
-     ::ident         ident                                  ; only for component-targeted loads
-     ::field         field                                  ; for component-targeted load
-     ::query         query'                                 ; query, relative to root of db OR component
-     ::post-mutation post-mutation
-     ::refresh       refresh
-     ::marker        marker
-     ::parallel      parallel
-     ::fallback      fallback}))
+    {::type                 :ready
+     ::uuid                 (uuid/uuid-string (uuid/make-random-squuid))
+     ::target               target
+     ::ident                ident                           ; only for component-targeted loads
+     ::field                field                           ; for component-targeted load
+     ::query                query'                          ; query, relative to root of db OR component
+     ::post-mutation        post-mutation
+     ::post-mutation-params post-mutation-params
+     ::refresh              refresh
+     ::marker               marker
+     ::parallel             parallel
+     ::fallback             fallback}))
 
 (defn mark-ready
   "Place a ready-to-load marker into the application state. This should be done from
   a mutate function that is abstractly loading something. This is intended for internal use.
 
-  See `load-field` for public API."
-  [& {:keys [state query ident field without params post-mutation fallback parallel refresh marker] :or {marker true refresh [] without #{}}}]
-  (swap! state update :om.next/ready-to-load conj
-         (ready-state
-           :ident ident
-           :field field
-           :parallel parallel
-           :refresh refresh
-           :marker marker
-           :params params
-           :without without
-           :query query
-           :post-mutation post-mutation
-           :fallback fallback)))
+  See the `load-data` and `load-field` functions in `untangled.client.data-fetch` for the public API."
+  [{:keys [state] :as config}]
+  (swap! state update :om.next/ready-to-load (fnil conj []) (ready-state (merge {:marker true :refresh [] :without #{}} config))))
 
-;; TODO: Rename "getters"
+(defn data-target
+  "Return the ident (if any) of the component related to the query in the data state marker. An ident is required
+  to be present if the marker is targeting a field."
+  [state] (::target state))
 (defn data-ident
   "Return the ident (if any) of the component related to the query in the data state marker. An ident is required
   to be present if the marker is targeting a field."
@@ -278,18 +273,19 @@
   "Get the 'primary' query key of the data fetch. This is defined as the first keyword of the overall query (which might
   be a simple prop or join key for example)"
   [state]
-  (let [expr (-> state ::query first)
-        key (cond
-              (keyword? expr) expr
-              (map? expr) (ffirst expr)
-              (list? expr) (ffirst (first expr)))]
-    key))
+  (let [ast (om/query->ast (-> state ::query))
+        node (-> ast :children first)]
+    (:key node)))
 
 (defn data-path
   "Get the app-state database path of the target of the load that the given data state marker is trying to load."
-  [state] (if (and (nil? (data-ident state)) (nil? (data-field state)))
-                          [(data-query-key state)]
-                          (conj (data-ident state) (data-field state))))
+  [state]
+  (let [target (data-target state)]
+    (cond
+      (and (vector? target) (not-empty target)) target
+      (and (vector? (data-ident state)) (keyword? (data-field state))) (conj (data-ident state) (data-field state))
+      :otherwise [(data-query-key state)])))
+
 (defn data-params
   "Get the parameters that the user wants to add to the first join/keyword of the data fetch query."
   [state] (::params state))
@@ -314,7 +310,8 @@
   (defn set-failed!
     "Returns a marker (based on the input state) that is in the error state"
     ([state] (set-failed! state nil))
-    ([state params] (set-type state :failed params))))
+    ([state params]
+     (set-type state :failed params))))
 
 (defn full-query
   "Composes together the queries of a sequence of data states into a single query."
@@ -325,6 +322,23 @@
   (let [state-atom (om/app-state reconciler)
         loading? (boolean (seq (get @state-atom :untangled/loads-in-progress)))]
     (swap! state-atom assoc :ui/loading-data loading?)))
+
+(defn relocate-targeted-results
+  "For items that are manually targeted, move them in app state from their result location to their target location."
+  [state-atom items]
+  (doseq [item items]
+    (let [default-target [(data-query-key item)]
+          field-target (conj (or (data-ident item) []) (::field item))
+          explicit-target (or (::target item) [])
+          relocate? (and (not-empty explicit-target)
+                         (not= explicit-target field-target)
+                         (not= explicit-target default-target))]
+      (when relocate?
+        (let [value (get-in @state-atom default-target)]
+          (swap! state-atom (fn [m]
+                              (-> m
+                                  (dissoc (data-query-key item))
+                                  (assoc-in explicit-target value)))))))))
 
 (defn- loaded-callback
   "Generates a callback that processes all of the post-processing steps once a remote load has completed. This includes:
@@ -348,18 +362,20 @@
           ran-mutations (atom false)
           remove-markers (fn [] (doseq [item loading-items]
                                   (swap! app-state (fn [s]
-                                                     (-> s
-                                                         (update :untangled/loads-in-progress disj (data-uuid item))
-                                                         (assoc-in (data-path item) nil))))))
+                                                     (cond-> s
+                                                             :always (update :untangled/loads-in-progress disj (data-uuid item))
+                                                             (data-marker? item) (assoc-in (data-path item) nil))))))
           run-post-mutations (fn [] (doseq [item loading-items]
                                       (when-let [mutation-symbol (::post-mutation item)]
                                         (reset! ran-mutations true)
-                                        (some->
-                                          (m/mutate {:state (om/app-state reconciler)} mutation-symbol {})
-                                          :action
-                                          (apply [])))))]
+                                        (let [params (or (::post-mutation-params item) {})]
+                                          (some->
+                                           (m/mutate {:state (om/app-state reconciler)} mutation-symbol params)
+                                           :action
+                                           (apply []))))))]
       (remove-markers)
       (om/merge! reconciler marked-response query)
+      (relocate-targeted-results app-state loading-items)
       (run-post-mutations)
       (set-global-loading reconciler)
       (if (or @ran-mutations (contains? refresh-set :untangled/force-root))
@@ -388,8 +404,8 @@
                         (doseq [item loading-items]
                           (swap! app-state (fn [s]
                                              (cond-> s
-                                               (data-marker? item) (update-in (data-path item) set-failed! error)
-                                               :always (update :untangled/loads-in-progress disj (data-uuid item)))))))
+                                                     (data-marker? item) (update-in (conj (data-path item) :ui/fetch-state) set-failed! error)
+                                                     :always (update :untangled/loads-in-progress disj (data-uuid item)))))))
           run-fallbacks (fn [] (doseq [item loading-items]
                                  (when-let [fallback-symbol (::fallback item)]
                                    (reset! ran-fallbacks true)